Convert storage-publiclink to service tracing
This converts the storage-publiclink service to the service tracing
methods.
